make len a size_t to fix problem on 64 bit system, reported by sebnod` in IRC
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@7643 06827809-a52a-0410-b366-d66718629ded

+++ agent/helpers/table.c
@@ -634,7 +634,7 @@ netsnmp_check_getnext_reply(netsnmp_request_info *request,
                             netsnmp_variable_list ** outvar)
 {
     oid      myname[MAX_OID_LEN];
-    int      myname_len;
+    size_t   myname_len;
 
     build_oid_noalloc(myname, MAX_OID_LEN, &myname_len,
                       prefix, prefix_len, newvar);
